+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 
 ### Bug fixes
 
+* Set bugsnag.span.category to 'custom' for custom spans.
+  [336](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/336)
+
 * Fix visionOS compilation errors. Note that visionOS is not yet officially supported.
   [327](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/327)